Refactoring of the as.spectral function to make better use of the built in class handling and generics. Documentation also becomes much easier, as the specific case can be described. The old documentation mainly described one case, where three was said to be possible.
I have not changed the as.spectral.trackdata code, only lifted it out into its own function. I am not sure it is correct though (as a result of the issue described above I belive) but I will look at it separately. as.spectral.vector is also not implemented yet (and I think it was really before either).