Lilly’s first version of serve() with Rda support
Only slightly edity by myself/RStudio.

emuR has had some support for signal tracks stored in Rda files since
version 2.4.2 (late 2023). The EMU-webApp has never had any support for
this. Now, the cleanest version in some sense would be to implement a
browser-side Rda parser; but since Rda is a complex format (especially
when compared to SSFF), that would be error-prone and the maintenance
burden would be way too high. See for example [1] for a glimpse of Rda’s
complexity or [2] for an incomplete and unmaintained attempt at an Rda
parser written in Javascript.

So the basic idea here is to have R load the Rda file just like in
get_trackdata() and convert it to SSFF on-the-fly whenever the browser
needs the data.

I outlined the idea and Lilly made this implementation.

[1] https://www.loc.gov/preservation/digital/formats/fdd/fdd000470.shtml
[2] https://github.com/zackorndorff/rdaparser