Fixing broken example for very old function `buildtrack`
I have had to correct this example at CRAN’s request, because it produces an
error with R-devel; probably starting with R revision r89098 from 2025-12-04.

Before that, this example produced no error; however, it did produce nonsense
results. I am not sure if this example ever worked as intended, or if a
regression was introduced at some point.

Today, I was able to confirm that it produces nonsense both in R 4.5.2 (which
is the current version) and in R 3.1.0 from 2014-04-10 (which is the earliest
version I can easily test with the Docker image rocker/r-ver).

What is the problem with the following example?

```
p = by(vowlax.fdat[1,2],INDICES=NULL, diff)
m = buildtrack(p)
```

Here, `p` is supposed to show the pairwise difference between adjacent
formant values; in fact, however, it is a data frame with 18 rows but no
columns. The subsetting [1,2] shows that it is supposed to work on the
second formant (F2) of  the first segment in the trackdata object
vowlax.fdat. The first segment is 18 samples long, so really we would
expect to see 17 pairwise differences of F2 values and not 18.

To make this example work as (I think it was) intended, I had to change
it to:

```
p = by(vowlax.fdat[1,2],INDICES=NULL, apply, 2, diff)
m = buildtrack(p)
```

This way, `p[[1]]` actually contains a 17x1 matrix with rownames according to
the timestamps in the original trackdata object (`vowlax.fdat$ftime`). Hence,
`buildtrack` can store a sensible value in `m`.

Using `sapply, diff` instead of `apply, 2, diff` also works for `by()`, but it
loses the timestamp information, so `buildtrack` does not work.

Was there a regression?

A variant of this example was already included in commit 45e382a5 from 2006,
which introduced buildtrack.S (right, S, not R!). I suspect that there was no
regression, but the example was not properly translated from S to R, or it was
never correct in the first place.